Done add cart & history food

+++ DATN/DATN/Client/Pages/OrderList.razor.cs
@@ -2,6 +2,7 @@
 using DATN.Shared;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Microsoft.JSInterop;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -12,23 +13,32 @@ namespace DATN.Client.Pages
     {
         private List<Cart> cartItems = new List<Cart>();
         private decimal Total;
-        private decimal Fee = 0;
         private decimal Pay;
 
         protected override async Task OnInitializedAsync()
         {
             cartItems = await _cartService.GetCartAsync();
+            CalculateTotal();
         }
 
         private async Task RemoveFromCart(Cart product)
         {
             if (product != null)
             {
-                await _cartService.RemoveItemFromCartAsync(product);
+                if (product.Quantity > 1)
+                {
+                    product.Quantity -= 1;
+                    await _cartService.SaveCartAsync(cartItems);
+                }
+                else
+                {
+                    await _cartService.RemoveItemFromCartAsync(product);
+                }
                 cartItems = await _cartService.GetCartAsync();
                 CalculateTotal();
             }
         }
+
         private async Task ThanhtoanAsync()
         {
             if (cartItems.Count() > 0)
@@ -37,7 +47,7 @@ namespace DATN.Client.Pages
             }
             else
             {
-                await JS.InvokeVoidAsync("showAlert", "EmptyPro");
+                await JS.InvokeVoidAsync("showAlert", "warning","Vui lòng thêm món ăn","");
             }
         }
         private void CalculateTotal()
@@ -45,9 +55,17 @@ namespace DATN.Client.Pages
             Total = 0;
             foreach (var item in cartItems)
             {
-                Total += item.Price;
+                Total += item.Price * item.Quantity;
             }
-            Pay = Total + Fee;
+        }
+        private async Task Order()
+        {
+            var expiryTime = DateTime.Now.AddMinutes(30).ToString("o");
+            await _localStorageService.SetCartItemAsync("historyOrder", cartItems);
+            await _localStorageService.SetItemAsync("cartExpiryTime", expiryTime);
+            await JS.InvokeVoidAsync("showAlert", "success","Đặt món thành công","Bạn vui lòng đợi đầu bếp làm nha :3");
+            await Task.Delay(1000);
+            Navigation.NavigateTo("/food-ordered");
         }
     }
 }
